+++ apps/main-app/ts/features/itwallet/presentation/details/screens/ItwPresentationCredentialDetailScreen.tsx
@@ -16,7 +16,6 @@ import {
   useIONavigation
 } from "../../../../../navigation/params/AppParamsList.ts";
 import { useIODispatch, useIOSelector } from "../../../../../store/hooks.ts";
-import { usePreventScreenCapture } from "../../../../../utils/hooks/usePreventScreenCapture.ts";
 import { identificationRequest } from "../../../../identification/store/actions";
 import { trackCredentialRenewStart } from "../../../analytics";
 import { getMixPanelCredential } from "../../../analytics/utils";
@@ -47,6 +46,7 @@ import { ITW_ROUTES } from "../../../navigation/routes.ts";
 import { ItwCredentialTrustmark } from "../../../trustmark/components/ItwCredentialTrustmark.tsx";
 import { trackItwProximityShowQrCode } from "../../proximity/analytics";
 import { ITW_PROXIMITY_ROUTES } from "../../proximity/navigation/routes";
+import { isPresentableCredentialSelector } from "../../proximity/store/selectors/credentials";
 import {
   trackCredentialDetail,
   trackWalletCredentialShowFAC_SIMILE,
@@ -66,6 +66,7 @@ import {
   CredentialCtaProps,
   ItwPresentationDetailsScreenBase
 } from "../components/ItwPresentationDetailsScreenBase.tsx";
+import { useItwDisplayCredentialStatus } from "../hooks/useItwDisplayCredentialStatus.tsx";
 import { shouldShowMdlUpdateDigitalCredential } from "../utils";
 
 export type ItwPresentationCredentialDetailNavigationParams = {
@@ -195,6 +196,10 @@ export const ItwPresentationCredentialDetail = ({
   const { status = "valid" } = useIOSelector(state =>
     itwCredentialStatusSelector(state, credential.credentialType)
   );
+  const isPresentableCredential = useIOSelector(
+    isPresentableCredentialSelector(credential.credentialType)
+  );
+  const displayStatus = useItwDisplayCredentialStatus(status);
   const contentClaim = credential.parsedCredential[WellKnownClaim.content];
   const hasSkeumorphicCard = credentialsWithSkeumorphicCard.includes(
     credential.credentialType
@@ -212,7 +217,7 @@ export const ItwPresentationCredentialDetail = ({
   );
 
   useDebugInfo(credential);
-  usePreventScreenCapture();
+  // TODO: [SIW-4622] re-enable usePreventScreenCapture();
 
   useFocusEffect(
     useCallback(() => {
@@ -277,10 +282,7 @@ export const ItwPresentationCredentialDetail = ({
       };
     }
 
-    if (
-      credentialType === CredentialType.DRIVING_LICENSE &&
-      itwFeaturesEnabled
-    ) {
+    if (itwFeaturesEnabled && isPresentableCredential) {
       return {
         label: I18n.t("features.itWallet.presentation.ctas.present"),
         icon: "productITWallet",
@@ -322,6 +324,7 @@ export const ItwPresentationCredentialDetail = ({
     shouldShowMdlUpdateCta,
     itwFeaturesEnabled,
     isL3Credential,
+    isPresentableCredential,
     contentClaim,
     navigation,
     mixPanelCredential,
@@ -343,8 +346,11 @@ export const ItwPresentationCredentialDetail = ({
       });
     } else {
       navigation.navigate(ITW_ROUTES.MAIN, {
-        screen: ITW_ROUTES.PRESENTATION.CREDENTIAL_CARD_SCREEN,
-        params: { credentialType: credential.credentialType }
+        screen: ITW_ROUTES.PRESENTATION.CREDENTIAL_CARD_MODAL,
+        params: {
+          credential,
+          status: displayStatus
+        }
       });
     }
   };