+++ apps/main-app/ts/features/messages/store/reducers/__tests__/messagePrecondition.test.ts
@@ -130,7 +130,6 @@ const computeExpectedOutput = (
     | typeof shownPreconditionStatusAction
     | typeof updateRequiredPreconditionStatusAction
   >
-  // eslint-disable-next-line complexity
 ) => {
   switch (fromStatus.state) {
     case "error":