+++ apps/main-app/ts/features/payments/checkout/analytics/index.ts
@@ -38,7 +38,6 @@ const MYBANK_PSP_BANNER_ID = "mybank_psp_selection";
 
 const getPaymentAnalyticsEventFromFailureOutcome = (
   outcome: WalletPaymentOutcomeEnum
-  // eslint-disable-next-line complexity
 ) => {
   switch (outcome) {
     case WalletPaymentOutcomeEnum.AUTH_ERROR: