+++ apps/main-app/ts/features/services/institution/hooks/useServicesFetcher.tsx
@@ -11,7 +11,7 @@ import {
   paginatedServicesSelector
 } from "../store/selectors";
 
-const LIMIT: number = 20;
+const LIMIT = 20;
 
 export const useServicesFetcher = (institutionId: string) => {
   const dispatch = useIODispatch();