+++ apps/main-app/ts/utils/analytics.ts
@@ -66,10 +66,7 @@ export const booleanToYesNo = (value: boolean): "yes" | "no" =>
     )
   );
 
-export const numberToYesNoOnThreshold = (
-  value: number,
-  threshold: number = 0
-) =>
+export const numberToYesNoOnThreshold = (value: number, threshold = 0) =>
   pipe(
     value > threshold,
     B.fold(